29/05/2025 Se realizara cambio en el sistema
Actualiza configuración y mejora la estructura del código

Se cambió la URL del servidor a `https://localhost` y se ajustó el puerto.
Se descomentaron líneas en `App.xaml.cs` para inicializar el servicio de autenticación.
Se añadieron nuevos paquetes en `Cliente.csproj` y se reorganizó la estructura de archivos.
Se implementó la interfaz `IDeleteable` en varios modelos y se introdujeron nuevos atributos personalizados.
Se añadieron clases de servicio y controlador para manejar CRUD de nuevas entidades.
Se mejoró la gestión de errores con la clase `ErrorResponse` y se optimizó la interfaz de usuario con nuevos diálogos.
Se implementaron métodos para inicializar datos por defecto y se mejoró la lógica de navegación en las vistas.

+++ Servidor/src/Repositorios/RepositorioBase.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 
 namespace Servidor.src.Repositorios
 {
-    public abstract class RepositorioBase<TObj> : IRepository<TObj> where TObj : IIdentifiable
+    public abstract class RepositorioBase<TObj> : IRepository<TObj> where TObj : IIdentifiable, IDeleteable
     {
         public IMongoCollection<TObj> Collection { get; }
 
@@ -108,6 +108,9 @@ namespace Servidor.src.Repositorios
         {
             try
             {
+                var obj = await GetByIdAsync(id);
+                if (obj == null || !obj.Deleteable) return false; // Si el objeto no existe, no se puede eliminar
+
                 var result = await Collection.DeleteOneAsync(x => x.Id == id);
                 return result.DeletedCount > 0;
             }
